requirements.txt
# 后端依赖
fastapi==0.104.1
uvicorn==0.24.0
psycopg2-binary==2.9.9
sqlalchemy==2.0.23
pydantic==2.5.0
python-dotenv==1.0.0
openai==1.3.7
requests==2.31.0
pandas==2.1.4
numpy==1.24.3
scikit-learn==1.3.2
python-multipart==0.0.6
jinja2==3.1.2
aiofiles==23.2.1

# 数据处理
regex==2023.10.3
jieba==0.42.1